+++ src/model/mod.rs
@@ -10,7 +10,3 @@ mod loan;
 pub use self::loan_builder::LoanBuilder;
 mod loan_builder;
 
-pub use self::resource::Resource;
-pub use self::resource::AccountInfoResource;
-pub use self::resource::LoansInfoResource;
-mod resource;
