+++ src/resource/mod.rs
@@ -1,0 +1,11 @@
+pub use self::resource_loader::ResourceLoader;
+mod resource_loader;
+
+pub use self::resource::Resource;
+mod resource;
+
+pub use self::account_resource::AccountInfoResource;
+mod account_resource;
+
+pub use self::loans_resource::LoansInfoResource;
+mod loans_resource;
