--- src/sync/mod.rs
@@ -1,2 +1,0 @@
-pub use self::resource_loader::ResourceLoader;
-mod resource_loader;
